pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prestamos-backend"
version = "0.1.0"
description = "SaaS backend para gestión de préstamos personales con validación OCR"
requires-python = ">=3.11"
authors = [
    { name = "TFG", email = "tfg@example.com" }
]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "email-validator>=2.2.0",
    "pydantic-settings>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "aioboto3>=13.0.0",
    "pillow>=10.3.0",
    "paddleocr==3.4.0",
    "numpy<2.0",
    "opencv-python-headless>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "ruff>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
